+++ src/app/signup/page.tsx
@@ -1,5 +1,4 @@
 'use client'
-
 import React, { useState, useRef, useEffect } from 'react'
 import { useRouter } from 'next/navigation'
 import Select from '@/components/signup/Select'
@@ -11,364 +10,26 @@ import EmailVerification from '@/components/common/EmailVerification'
 import Link from 'next/link'
 import Lottie from 'lottie-react'
 import loading from '../../../public/loading.json'
+import { useSignupHandler } from '@/hooks/auth/useSignupHandler'
 
 const Signup = () => {
-  const [signupError, setSignupError] = useState<string>('')
-  const router = useRouter()
-
-  const [step, setStep] = useState(1)
-  const [isLoading, setIsLoading] = useState(false)
-  const [formData, setFormData] = useState({
-    name: '',
-    email: '',
-    password: '',
-    confirmPassword: '',
-    isVerified: false,
-
-    school: '',
-    customSchool: '', // “해당 없음” 선택 시, 직접 입력한 학교명
-    classYear: '',
-    selectedBatch: '',
-    resumeTitle: '',
-    githubUrl: '',
-    tistoryUrl: '',
-    mediumUrl: '',
-    velogUrl: '',
-    selectedPositions: [] as string[],
-    recommendation: null as string | null,
-    employmentStatus: null as string | null,
-    // CareerToggle에서 "있어요" 버튼을 누르면 'yes'로 저장됩니다.
-    internshipExperience: null as string | null,
-    jobExperience: null as string | null,
-
-    // 경험 정보: 기존의 키 대신 unified한 키로 관리
-    internships: [] as {
-      companyName: string
-      position: string
-      startDate: string
-      endDate: string
-      isCurrentJob: boolean
-    }[],
-    fullTimes: [] as {
-      companyName: string
-      position: string
-      startDate: string
-      endDate: string
-      isCurrentJob: boolean
-    }[],
-
-    // 이력서 추가 정보
-    resumeFile: null as File | null,
-    resumeCategory: 'RESUME',
-    resumeIsMain: true,
-    resumePosition: '',
-  })
-
-  const handleChange = (
-    e:
-      | React.ChangeEvent<HTMLInputElement>
-      | React.ChangeEvent<HTMLSelectElement>,
-  ) => {
-    const { name, value } = e.target
-    setFormData((prev) => ({
-      ...prev,
-      [name]: value,
-    }))
-  }
-
-  const passwordsMatch =
-    formData.password === formData.confirmPassword &&
-    formData.password.length > 0
-
-  const handleNext = () => {
-    setSignupError('')
-    // 스텝 1 필드 검증: 각 필드가 비어있거나, isVerified가 false면 alert 후 진행하지 않음
-    if (!formData.name.trim()) {
-      alert('이름을 입력해주세요.')
-      return
-    }
-    if (!formData.email.trim()) {
-      alert('이메일을 입력해주세요.')
-      return
-    }
-    if (!formData.password) {
-      alert('비밀번호를 입력해주세요.')
-      return
-    }
-    const passwordRegex = /^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&]).{8,}$/
-    if (!passwordRegex.test(formData.password)) {
-      alert(
-        '비밀번호 양식에 맞지 않습니다.\n 영어, 숫자, 특수문자를 각각 최소 한 개 이상 포함해주세요.\n 비밀번호는 8자 이상이어야 합니다.',
-      )
-      return
-    }
-    if (!formData.confirmPassword) {
-      alert('비밀번호 확인을 입력해주세요.')
-      return
-    }
-    if (!formData.isVerified) {
-      alert('이메일 인증을 완료해주세요.')
-      return
-    }
-    if (!passwordsMatch) {
-      alert('비밀번호가 일치하지 않습니다.')
-      return
-    }
-    // 모든 필드가 유효하면 스텝 진행
-    if (step < 2) {
-      setStep(step + 1)
-    }
-  }
-
-  const handleBack = () => {
-    setSignupError('')
-    if (step > 1) {
-      setStep(step - 1)
-    }
-  }
-
-  const handleSignup = async () => {
-    setSignupError('')
-
-    if (!formData.school.trim()) {
-      alert('대학교를 선택해주세요.')
-      return
-    }
-    if (formData.school === '해당 없음' && !formData.customSchool.trim()) {
-      alert('학교명을 직접 입력해주세요.')
-      return
-    }
-    if (!formData.classYear.trim()) {
-      alert('학년을 선택해주세요.')
-      return
-    }
-    if (!formData.selectedBatch.trim()) {
-      alert('기수를 선택해주세요.')
-      return
-    }
-    if (!formData.resumeTitle.trim()) {
-      alert('이력서 제목을 입력해주세요.')
-      return
-    }
-    if (!formData.resumeFile) {
-      alert('이력서 파일을 업로드해주세요.')
-      return
-    }
-    if (!formData.resumeCategory.trim()) {
-      alert('이력서 카테고리를 선택해주세요.')
-      return
-    }
-    if (!formData.resumePosition) {
-      alert('이력서 포지션을 선택해주세요.')
-      return
-    }
-    if (formData.selectedPositions.length === 0) {
-      alert('포지션을 선택해주세요.')
-      return
-    }
-    if (!formData.recommendation) {
-      alert('추천 여부를 선택해주세요.')
-      return
-    }
-    if (!formData.githubUrl.trim()) {
-      alert('깃허브 URL을 입력해주세요.')
-      return
-    }
-
-    if (!formData.employmentStatus) {
-      alert('경력 여부를 선택해주세요.')
-      return
-    }
-    if (formData.employmentStatus === 'yes') {
-      if (!formData.internshipExperience) {
-        alert('인턴 경험 여부를 선택해주세요.')
-        return
-      }
-      if (!formData.jobExperience) {
-        alert('정규직 경험 여부를 선택해주세요.')
-        return
-      }
-    }
-    if (!formData.isVerified) {
-      alert('이메일 인증을 완료해주세요.')
-      return
-    }
-
-    // 모든 검증 통과 후 로딩 시작
-    setIsLoading(true)
-
-    // 경험 데이터를 백엔드 형식으로 통합
-    const experiences: {
-      companyName: string
-      position: string
-      startDate: string
-      endDate: string
-      category: string
-    }[] = []
-
-    if (formData.internshipExperience === 'yes') {
-      experiences.push(
-        ...formData.internships.map((item) => ({
-          companyName: item.companyName || '',
-          position: item.position || formData.selectedPositions[0] || '',
-          startDate: item.startDate || '',
-          endDate: item.isCurrentJob ? '' : item.endDate || '',
-          category: '인턴',
-        })),
-      )
-    }
-
-    if (formData.jobExperience === 'yes') {
-      experiences.push(
-        ...formData.fullTimes.map((item) => ({
-          companyName: item.companyName || '',
-          position: item.position || formData.selectedPositions[0] || '',
-          startDate: item.startDate || '',
-          endDate: item.isCurrentJob ? '' : item.endDate || '',
-          category: '정규직',
-        })),
-      )
-    }
-
-    const finalSchool =
-      formData.school === '해당 없음'
-        ? formData.customSchool.trim()
-        : formData.school
-
-    const createUserRequest: any = {
-      mainPosition: formData.selectedPositions[0] || '',
-      subPosition: formData.selectedPositions[1] || '',
-      name: formData.name,
-      githubUrl: formData.githubUrl,
-      isLft: formData.recommendation === 'yes',
-      school: finalSchool,
-      grade: formData.classYear,
-      password: formData.password,
-      email: formData.email,
-      year: parseInt(formData.selectedBatch),
-    }
-
-    if (formData.tistoryUrl) {
-      createUserRequest.tistoryUrl = formData.tistoryUrl
-    }
-    if (formData.mediumUrl) {
-      createUserRequest.mediumUrl = formData.mediumUrl
-    }
-    if (formData.velogUrl) {
-      createUserRequest.velogUrl = formData.velogUrl
-    }
-
-    const requestPayload = {
-      createUserRequest,
-      createUserExperienceRequest: {
-        experiences,
-      },
-      createResumeRequest: {
-        category: formData.resumeCategory,
-        position: formData.resumePosition,
-        title: formData.resumeTitle,
-        isMain: formData.resumeIsMain,
-      },
-    }
-
-    try {
-      const formDataToSend = new FormData()
-      if (formData.resumeFile) {
-        formDataToSend.append('file', formData.resumeFile)
-      }
-      formDataToSend.append(
-        'createUserWithResumeRequest',
-        JSON.stringify(requestPayload),
-      )
-
-      const response = await fetch('/api/v1/users/signup', {
-        method: 'POST',
-        body: formDataToSend,
-      })
-
-      if (!response.ok) {
-        const errorData = await response.json().catch(() => null)
-
-        if (response.status === 400) {
-          alert('필수 항목을 모두 입력해주세요.')
-          return
-        }
-
-        if (response.status === 500) {
-          alert('이미 등록된 이메일입니다.')
-          return
-        }
-
-        if (response.status === 401) {
-          alert(errorData.message)
-          return
-        }
-      }
-
-      router.push('/login?form=signup')
-    } catch (err: any) {
-      alert('네트워크 오류가 발생했습니다.')
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  const handlePositionSelect = (position: string) => {
-    if (formData.selectedPositions.includes(position)) {
-      setFormData((prev) => ({
-        ...prev,
-        selectedPositions: prev.selectedPositions.filter(
-          (item) => item !== position,
-        ),
-      }))
-    } else if (formData.selectedPositions.length < 2) {
-      setFormData((prev) => ({
-        ...prev,
-        selectedPositions: [...prev.selectedPositions, position],
-      }))
-    }
-  }
-
-  const setInternships = (internships: any[]) => {
-    setFormData((prev) => ({
-      ...prev,
-      internships,
-    }))
-  }
-
-  const setFullTimes = (fullTimes: any[]) => {
-    setFormData((prev) => ({
-      ...prev,
-      fullTimes,
-    }))
-  }
-
-  const formRef = useRef<HTMLDivElement>(null)
-  const bottomRef = useRef<HTMLDivElement>(null)
-
-  useEffect(() => {
-    const formElement = formRef.current
-    if (formElement && bottomRef.current) {
-      if (
-        formData.employmentStatus === 'yes' ||
-        formData.internshipExperience === 'yes' ||
-        formData.jobExperience === 'yes' ||
-        formData.internships.length > 0 ||
-        formData.fullTimes.length > 0
-      ) {
-        formElement.scrollTo({
-          top: bottomRef.current.offsetTop,
-          behavior: 'smooth',
-        })
-      }
-    }
-  }, [
-    formData.employmentStatus,
-    formData.internshipExperience,
-    formData.jobExperience,
-    formData.internships.length,
-    formData.fullTimes.length,
-  ])
+  const {
+    formData,
+    setFormData,
+    setInternships,
+    setFullTimes,
+    handleChange,
+    handlePositionSelect,
+    step,
+    setStep,
+    isLoading,
+    handleSignup,
+    passwordsMatch,
+    formRef,
+    bottomRef,
+    handleNext,
+    handleBack,
+  } = useSignupHandler()
 
   return (
     <div className="flex min-h-[calc(100vh-61px)] items-center justify-between">
